src/my_model/MyBaseLayer/requirement.txt
happybase==1.2.0
numpy==1.16.4
pandas==0.24.2
ply==3.11
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
thriftpy2==0.4.4
